Verilog.jl

Documentation for Verilog.

Verilog.RelativeRange

Verilog.VerilogRange

Verilog.Wire

Verilog.msb

Verilog.v

Verilog.find_dependencies

Verilog.generate_verilog_file

Verilog.verilate

Verilog.verilator_adapter

Verilog.@input

Verilog.@verilate

Verilog.RelativeRange — Type

RelativeRange

is a range object that can take values which can be relative to the most significant bit.

Verilog.VerilogRange — Type

VerilogRange

is a range object that represents a verilog range. These are declared by using the v ranges, so 5:0v is a VerilogRange from 0 to 5.

Verilog.Wire — Type

Wire{R}

is the basic type for Verilog operations. R specifies a "unit range" of integers. Use the "v" suffix to enable verilog-style ranging.

Wire{3:0v} declares a four-digit verilog wire with indices spanning from 0->3; Wire{6:2v} declares a five-bit verilog wire with indices spanning from 2->6.

Verilog.msb — Type

msb

is a special keyword which indicates that the value should go to the most significant bit. There is no equivalent "lsb", since in most cases this will be a fairly easy-to-reference value. Plan your code accordingly.

For example: mywire[16:0v] = Wire(0xF000, 16) mywire[msb:12] # ==> Wire{3:0v}(0xF)

mywire2[16:0v] = Wire(0x0F00, 16) my_wire[(msb-4):(msb-8)] # ==> Wire{3:0v}(0xF)

Verilog.v — Type

v

is an appendage that reverses the order of julia's ranges so that they have the cosmetic appearance of a verilog range. This can be used both in setters and getters for wire collections.

For example: my_wire[5:0v] references a wire over indices 0-5 my_wire[6:2v] references a wire over indices 2-6

For dereferencing (getting) you can reverse this: my_wire[2:6v] references the wire over indices 2-6 with the msb order reversed.

if you are using a variable, be sure to put the value in paretheses.

For example:

my_wire[bits:(bits-4)v]

Verilog.find_dependencies — Function

Verilog.find_dependencies(mod, (params...))

outputs the set of dependencies of a certain module.

Verilog.generate_verilog_file — Method

generate_verilog_file(path, mod, (params...); exclude = [])

pass this function a path, (verilog module) name and the parameters, and it puts that module and its dependencies into a file in the desired path. The filename will be (module name + suffix).v, with module name & suffix as generated by the interal specifications.

Verilog.verilate — Method

verilate(mod::Function, p::Tuple) uses verilator to create a .so file corresponding to the library. Each of the functions in the mods array will be the top-level element in a freestanding .v file. They may refer to each other. the "mods" are a tuple of the function and the parameter.

Passed options: path::String - defaults to "." - the library .so file will be placed in the path. libname::String - file name for the library. Defaults to libVerilated-[functionwithsuffix].so with_source::Bool - defaults to "false" - should the c source be included? this is useful if you're building a library around the verilated module.

Verilog.verilator_adapter — Method

Verilog.verilator_adapter(id)

creates the required C file for verilation.

Verilog.@input — Macro

@input identifier rangedescriptor

binds an input to a given range descriptor.

Verilog.@verilate — Macro

@verilate function parameters [path]

verilates a @verilog function with the passed parameter and places the shared object library into [path]. By default, the path used is going to be "./libVerilated-[function].so".

the function is then hooked up to a function with signature []